+++ ProcessQueue/__init__.py
@@ -347,7 +347,7 @@ def reuse_existing_field_concepts(new_fields_map, content_type, api_url, headers
         ids_to_get = ",".join(map(str, ids))
 
         get_field_tables = requests.get(
-            url=f"{api_url}scanreportfieldsfilter/?id__in={ids_to_get}&fields=id,"
+            url=f"{api_url}scanreportfields/?id__in={ids_to_get}&fields=id,"
             f"scan_report_table",
             headers=headers,
         )
@@ -363,7 +363,7 @@ def reuse_existing_field_concepts(new_fields_map, content_type, api_url, headers
         ids_to_get = ",".join(map(str, ids))
 
         get_field_tables = requests.get(
-            url=f"{api_url}scanreportfieldsfilter/?name__in={ids_to_get}&fields=id,"
+            url=f"{api_url}scanreportfields/?name__in={ids_to_get}&fields=id,"
             f"scan_report_table",
             headers=headers,
         )
@@ -392,7 +392,7 @@ def reuse_existing_field_concepts(new_fields_map, content_type, api_url, headers
         ids_to_get = ",".join(map(str, ids))
 
         get_field_tables = requests.get(
-            url=f"{api_url}scanreporttablesfilter/?id__in={ids_to_get}&fields=id,"
+            url=f"{api_url}scanreporttables/?id__in={ids_to_get}&fields=id,"
             f"scan_report",
             headers=headers,
         )
@@ -407,14 +407,20 @@ def reuse_existing_field_concepts(new_fields_map, content_type, api_url, headers
         url=f"{api_url}scanreports/",
         headers=headers,
     )
+    get_datasets = requests.get(
+        url=f"{api_url}datasets/",
+        headers=headers,
+    )
     # get active scanreports and map them to fields. Remove any fields in archived
     # reports or not marked as 'Mapping Complete'
-    active_srs = [
-        str(item["id"])
-        for item in json.loads(get_scan_reports.content.decode("utf-8"))
-        if item["hidden"] is False and item["status"] == "COMPLET"
-    ]
-    # active reports is list of report ids that are not archived and have the status
+    active_srs = []
+    for item in get_scan_reports.json():
+        if item["hidden"] is False and item["status"] == "COMPLET":
+            for ds in get_datasets.json():
+                # Exclude scan reports if their parent_dataset is archived
+                if ds["id"] == item["parent_dataset"] and ds["hidden"] is False:
+                    active_srs.append(str(item["id"]))
+    # active reports is list of report ids that belong to an active dataset, are not archived, and have the status
     # 'Mapping Complete'
 
     # map value id to active scan report
@@ -552,7 +558,7 @@ def reuse_existing_value_concepts(new_values_map, content_type, api_url, headers
     for ids in new_paginated_field_ids:
         ids_to_get = ",".join(map(str, ids))
         get_fields = requests.get(
-            url=f"{api_url}scanreportfieldsfilter/?id__in={ids_to_get}&fields=id,name",
+            url=f"{api_url}scanreportfields/?id__in={ids_to_get}&fields=id,name",
             headers=headers,
         )
         new_fields.append(json.loads(get_fields.content.decode("utf-8")))
@@ -602,7 +608,7 @@ def reuse_existing_value_concepts(new_values_map, content_type, api_url, headers
         ids_to_get = ",".join(map(str, ids))
 
         get_field_tables = requests.get(
-            url=f"{api_url}scanreportvaluesfilter/?id__in={ids_to_get}&fields=id,value,scan_report_field,"
+            url=f"{api_url}scanreportvalues/?id__in={ids_to_get}&fields=id,value,scan_report_field,"
             f"value_description",
             headers=headers,
         )
@@ -621,7 +627,7 @@ def reuse_existing_value_concepts(new_values_map, content_type, api_url, headers
         new_values_names = ",".join(map(str, names))
 
         get_field_tables = requests.get(
-            url=f"{api_url}scanreportvaluesfilter/?value__in={new_values_names}&fields="
+            url=f"{api_url}scanreportvalues/?value__in={new_values_names}&fields="
             f"id,value,scan_report_field,value_description",
             headers=headers,
         )
@@ -658,7 +664,7 @@ def reuse_existing_value_concepts(new_values_map, content_type, api_url, headers
         ids_to_get = ",".join(map(str, ids))
 
         get_value_fields = requests.get(
-            url=f"{api_url}scanreportfieldsfilter/?id__in={ids_to_get}&fields=id,"
+            url=f"{api_url}scanreportfields/?id__in={ids_to_get}&fields=id,"
             f"name,scan_report_table",
             headers=headers,
         )
@@ -679,7 +685,7 @@ def reuse_existing_value_concepts(new_values_map, content_type, api_url, headers
         ids_to_get = ",".join(map(str, ids))
 
         get_field_tables = requests.get(
-            url=f"{api_url}scanreporttablesfilter/?id__in={ids_to_get}&fields=id,"
+            url=f"{api_url}scanreporttables/?id__in={ids_to_get}&fields=id,"
             f"scan_report",
             headers=headers,
         )
@@ -694,14 +700,20 @@ def reuse_existing_value_concepts(new_values_map, content_type, api_url, headers
         url=f"{api_url}scanreports/",
         headers=headers,
     )
+    get_datasets = requests.get(
+        url=f"{api_url}datasets/",
+        headers=headers,
+    )
     # get active scanreports and map them to fields. Remove any fields in archived
     # reports or not marked as 'Mapping Complete'
-    active_srs = [
-        str(item["id"])
-        for item in json.loads(get_scan_reports.content.decode("utf-8"))
-        if item["hidden"] is False and item["status"] == "COMPLET"
-    ]
-    # active reports is list of report ids that are not archived and have the status
+    active_srs = []
+    for item in get_scan_reports.json():
+        if item["hidden"] is False and item["status"] == "COMPLET":
+            for ds in get_datasets.json():
+                # Exclude scan reports if their parent_dataset is archived
+                if ds["id"] == item["parent_dataset"] and ds["hidden"] is False:
+                    active_srs.append(str(item["id"]))
+    # active reports is list of report ids that belong to an active dataset, are not archived, and have the status
     # 'Mapping Complete'
 
     # map value id to active scan report