+++ react-client-app/src/components/FieldsTbl.jsx
@@ -21,13 +21,16 @@ import {
 } from "@chakra-ui/react"
 
 import { Formik, Form, } from 'formik'
-import {getScanReportTable, getScanReportFieldValues} from '../api/values'
+import { getScanReportTable, getScanReportFieldValues, useGet } from '../api/values'
 import ConceptTag from './ConceptTag'
 import ToastAlert from './ToastAlert'
 
 const FieldsTbl = (props) => {
     // get the value to use to query the fields endpoint from the page url
-    const value = parseInt(new URLSearchParams(window.location.search).get("search"))
+    const pathArray = window.location.pathname.split("/")
+    const scanReportId = pathArray[pathArray.length - 3]
+    // const scanReportTableId = pathArray[pathArray.length - 1]
+    const scanReportTableId = window.pk ? window.pk : parseInt(new URLSearchParams(window.location.search).get("search"))
     const [alert, setAlert] = useState({ hidden: true, title: '', description: '', status: 'error' });
     const { isOpen, onOpen, onClose } = useDisclosure()
     const [values, setValues] = useState([]);
@@ -40,33 +43,46 @@ const FieldsTbl = (props) => {
 
     useEffect(() => {
         // run on initial render
-        // get field table values for specified id
-        getScanReportFieldValues(value, valuesRef).then(val => {
-            setValues(val)
-            setLoading(false)
-        })
-        // get scan report table data to use for checking person id and date event
-        getScanReportTable(value).then(table => {
-            scanReportTable.current = table
-            setMappingButtonDisabled(false)
-        })
+        // Check if user can see SR table
+        useGet(`/scanreporttables/${scanReportTableId}`).then(
+            res => {
+                // get field table values for specified id
+                getScanReportFieldValues(scanReportTableId, valuesRef).then(val => {
+                    setValues(val)
+                    setLoading(false)
+                })
+                // get scan report table data to use for checking person id and date event
+                getScanReportTable(scanReportTableId).then(table => {
+                    scanReportTable.current = table
+                    setMappingButtonDisabled(false)
+                })
+            }
+        ).catch(
+            err => {
+                // If user can't see SR table, show an error message
+                setError("Could not access the resource you requested. "
+                    + "Check that it exists and that you have permission to view it."
+                )
+                setLoading(false)
+            }
+        )
     }, []);
 
     // called to submit a concept to be added. Calls handle submit function from app.js
     const handleSubmit = (id, concept) => {
-        props.handleSubmit(id,concept,valuesRef,setValues,setAlert,onOpen,scanReportTable.current,15)
+        props.handleSubmit(id, concept, valuesRef, setValues, setAlert, onOpen, scanReportTable.current, 15)
     }
 
     // called to delete a concept. Calls handle delete function from app.js
     const handleDelete = (id, conceptId) => {
-        props.handleDelete(id,conceptId,valuesRef,setValues,setAlert,onOpen)
+        props.handleDelete(id, conceptId, valuesRef, setValues, setAlert, onOpen)
     }
 
     if (error) {
         //Render Error State
         return (
             <Flex padding="30px">
-                <Flex marginLeft="10px">An Error has occured while fetching values</Flex>
+                <Flex marginLeft="10px">{error}</Flex>
             </Flex>
         )
     }
@@ -91,10 +107,10 @@ const FieldsTbl = (props) => {
     else {
         return (
             <div>
-                {isOpen&&
-                <ScaleFade initialScale={0.9} in={isOpen}>
-                    <ToastAlert hide={onClose} title={alert.title} status={alert.status} description={alert.description} />
-                </ScaleFade>
+                {isOpen &&
+                    <ScaleFade initialScale={0.9} in={isOpen}>
+                        <ToastAlert hide={onClose} title={alert.title} status={alert.status} description={alert.description} />
+                    </ScaleFade>
                 }
 
                 <Table variant="striped" colorScheme="greyBasic">
@@ -106,7 +122,7 @@ const FieldsTbl = (props) => {
                             <Th>Data type</Th>
                             <Th>Concepts</Th>
                             <Th></Th>
-                            <Th>Edit</Th>
+                            {window.canEdit && <Th>Edit</Th>}
                         </Tr>
                     </Thead>
                     <Tbody>
@@ -114,7 +130,7 @@ const FieldsTbl = (props) => {
                             // Create new row for every value object
                             values.map((item, index) =>
                                 <Tr key={item.id}>
-                                    <Td><Link style={{ color: "#0000FF", }} href={"/values/?search=" + item.id}>{item.name}</Link></Td>
+                                    <Td><Link style={{ color: "#0000FF", }} href={`/scanreports/${scanReportId}/tables/${scanReportTableId}/fields/${item.id}`}>{item.name}</Link></Td>
                                     <Td maxW="250px"><Text maxW="100%" w="max-content">{item.description_column}</Text></Td>
                                     <Td>{item.type_column}</Td>
 
@@ -123,8 +139,15 @@ const FieldsTbl = (props) => {
                                             item.concepts.length > 0 &&
                                             <VStack alignItems='flex-start' >
                                                 {item.concepts.map((concept) => (
-                                                    <ConceptTag key={concept.concept.concept_id} conceptName={concept.concept.concept_name} conceptId={concept.concept.concept_id.toString()} conceptIdentifier={concept.id.toString()} itemId={item.id} handleDelete={handleDelete} 
-                                                    creation_type={concept.creation_type?concept.creation_type:undefined}/>
+                                                    <ConceptTag
+                                                        key={concept.concept.concept_id}
+                                                        conceptName={concept.concept.concept_name}
+                                                        conceptId={concept.concept.concept_id.toString()}
+                                                        conceptIdentifier={concept.id.toString()} itemId={item.id}
+                                                        handleDelete={handleDelete}
+                                                        creation_type={concept.creation_type ? concept.creation_type : undefined}
+                                                        readOnly={!window.canEdit}
+                                                    />
                                                 ))}
                                             </VStack>
                                             :
@@ -152,16 +175,18 @@ const FieldsTbl = (props) => {
                                                             name='concept'
                                                             value={values.concept}
                                                             onChange={handleChange}
-                                                            onBlur={handleBlur} />
+                                                            onBlur={handleBlur}
+                                                            isDisabled={!window.canEdit}
+                                                        />
                                                         <div>
-                                                            <Button type='submit' disabled={false} backgroundColor='#3C579E' color='white'>Add</Button>
+                                                            <Button type='submit' isDisabled={!window.canEdit} backgroundColor='#3C579E' color='white'>Add</Button>
                                                         </div>
                                                     </HStack>
                                                 </Form>
                                             )}
                                         </Formik>
                                     </Td>
-                                    <Td><Link style={{ color: "#0000FF", }} href={"/fields/" + item.id + "/update/"}>Edit Field</Link></Td>
+                                    {window.canEdit && <Td><Link style={{ color: "#0000FF", }} href={"/fields/" + item.id + "/update/"}>Edit Field</Link></Td>}
                                 </Tr>
                             )
                         }